+++ test/fzf-in-action.py
@@ -8,13 +8,13 @@ from pydymenu.demo_data import gen_options, list_options
 print("TESTING FZF")
 
 # Simple
-fzf = pydymenu.fzf(list_options)
+# fzf = pydymenu.fzf(list_options)
 
 # prompt
 # fzf = pydymenu.fzf(gen_options, prompt="Who can help guide the way? ")
 
 # # multi True
-# fzf = pydymenu.fzf(list_options, prompt="Who can help guide the way? ", multi=True)
+fzf = pydymenu.fzf(gen_options, prompt="Who can help guide the way? ", multi=True)
 
 # # multi False
 # fzf = pydymenu.fzf(gen_options, prompt="Who can help guide the way? ", multi=False)