+++ CHANGELOG.md
@@ -4,6 +4,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.5.0] - 2021-09-28
+
+### Changed
+
+- Big usability improvement. Both `rofi` and `fzf` now stream their output from 
+  a given iterable. This makes the user experience similar to `iterfzf`.
+
 ## [0.4.0] - 2021-09-21
 
 - Major code refactor and cleanup